split TIM4 files and add STM8S103 lib with debug information
